clip-image-search-engine/requirements.txt
torch==2.1.1
transformers==4.32.1
Pillow==9.4.0
pysolr==3.9.0
elasticsearch==7.13.4
faiss-cpu==1.7.4